Remove is_html_element_in_html_document
This change fixes some selectors for documents with no namespace set,
i.e. HTML (non-XML) where the namespace is not set in the html tag.

HTML doesn't officially support namespaces, but ElementTree sets the
information on tags anyway and we may care in the future if we want to handle
HTML mixed with XML (for example when SVG is directly embedded in the HTML
document).

Some selectors probably match too many elements now (SVG elementst in HTML
documents for sure), but it's still an improvement as it fixe simportant real
use cases (:link, selectors with uppercase characters, etc).

To definitely fix corner cases, I think that we should set the default HTML
namespace if none is set on HTML documents, as proposed in the
specification. This is of course left as an exercise for the reader.